Don't hardcode the server's hostname
This fixes a usability bug where the server is run inside of a Docker
container (and probably also when inside of VM?). Since the container is
run on a different IP the server wouldn't listen to any incoming
requests.

This does come with the downside that the server will now accept
connections on any IP address.